Make sprinkles global
